+++ src/models/kucoin_model.py
@@ -23,6 +23,6 @@ class KucoinRawData(BaseModel):
 
 class KucoinTransformedData(BaseModel):
     symbol: str
-    price: double
+    price: float
     time: datetime
     source: str
